# gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "US Government Data API",
    "description": "Unified REST API providing access to 45 US government data endpoints across 13 categories including cybersecurity vulnerability intelligence. Real-time data from NVD, CISA KEV, EPSS, NHTSA, FDA, CFPB, FEMA, NOAA, USGS, SEC, BLS, USDA, EPA, HUD, FAA, CBP, SAM.gov, CourtListener, CDC, WHO, Federal Register, Grants.gov, OpenCorporates, and more.",
    "version": "1.1.0",
    "contact": {
      "name": "MARTC03",
      "url": "https://apify.com/martc03"
    }
  },
  "servers": [
    {
      "url": "https://govdata-api.netlify.app",
      "description": "Production (Netlify Proxy)"
    }
  ],
  "paths": {
...
    "/api/v1/regulations/documents": {
      "get": {
        "operationId": "searchRegulations",
        "summary": "Search Federal Register Documents",
        "description": "Search Federal Register regulatory documents.",
        "tags": ["Regulations"],
        "parameters": [
          { "name": "term", "in": "query", "schema": { "type": "string" }, "description": "Search term" },
          { "name": "documentType", "in": "query", "schema": { "type": "string" }, "description": "Document type" },
          { "name": "agency", "in": "query", "schema": { "type": "string" }, "description": "Agency" },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date from" },
          { "name": "dateTo", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date to" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/regulations/document/{documentNumber}": {
      "get": {
        "operationId": "getRegDocument",
        "summary": "Get Federal Register Document",
        "description": "Get full details of a Federal Register document by number.",
        "tags": ["Regulations"],
        "parameters": [
          { "name": "documentNumber", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Federal Register document number" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/regulations/public-inspection": {
      "get": {
        "operationId": "searchPublicInspection",
        "summary": "Search Public Inspection Files",
        "description": "Search pre-publication Federal Register documents on public inspection.",
        "tags": ["Regulations"],
        "parameters": [
          { "name": "agency", "in": "query", "schema": { "type": "string" }, "description": "Agency" },
          { "name": "documentType", "in": "query", "schema": { "type": "string" }, "description": "Document type" },
          { "name": "postedFrom", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Posted from" },
          { "name": "postedTo", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Posted to" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/regulations/agencies": {
      "get": {
        "operationId": "listAgencies",
        "summary": "List Federal Agencies",
        "description": "List all federal agencies in the Federal Register.",
        "tags": ["Regulations"],
        "parameters": [],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
...
    "/api/v1/entities/companies": {
      "get": {
        "operationId": "searchCompanies",
        "summary": "Search Company Registry",
        "description": "Search OpenCorporates company registry.",
        "tags": ["Entities"],
        "parameters": [
          { "name": "name", "in": "query", "schema": { "type": "string" }, "description": "Company name" },
          { "name": "jurisdiction", "in": "query", "schema": { "type": "string" }, "description": "Jurisdiction" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" },
          { "name": "offset", "in": "query", "schema": { "type": "integer" }, "description": "Results offset" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/entities/company-details": {
      "get": {
        "operationId": "getCompanyDetails",
        "summary": "Get Company Details",
        "description": "Get company details including officers, filings, and history.",
        "tags": ["Entities"],
        "parameters": [
          { "name": "companyNumber", "in": "query", "schema": { "type": "string" }, "description": "Company number" },
          { "name": "jurisdiction", "in": "query", "schema": { "type": "string" }, "description": "Jurisdiction" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/entities/sec-companies": {
      "get": {
        "operationId": "searchSecCompanies",
        "summary": "Search SEC Company Filings",
        "description": "Search SEC EDGAR company filings.",
        "tags": ["Entities"],
        "parameters": [
          { "name": "searchText", "in": "query", "schema": { "type": "string" }, "description": "Search text" },
          { "name": "formTypes", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated form types" },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date from" },
          { "name": "dateTo", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date to" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/intel/company-filings": {
      "get": {
        "operationId": "getCompanyFilings",
        "summary": "Get Company SEC Filings",
        "description": "Get SEC filings for competitive analysis.",
        "tags": ["Intelligence"],
        "parameters": [
          { "name": "companyName", "in": "query", "schema": { "type": "string" }, "description": "Company name" },
          { "name": "formType", "in": "query", "schema": { "type": "string" }, "description": "SEC form type" },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date from" },
          { "name": "dateTo", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date to" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/intel/company-news": {
      "get": {
        "operationId": "getCompanyNews",
        "summary": "Get Company News",
        "description": "Get latest news for a company via Google News RSS.",
        "tags": ["Intelligence"],
        "parameters": [
          { "name": "companyName", "in": "query", "schema": { "type": "string" }, "description": "Company name" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/intel/company-contracts": {
      "get": {
        "operationId": "getCompanyContracts",
        "summary": "Get Company Government Contracts",
        "description": "Get USAspending government contracts for a company.",
        "tags": ["Intelligence"],
        "parameters": [
          { "name": "companyName", "in": "query", "schema": { "type": "string" }, "description": "Company name" },
          { "name": "awardType", "in": "query", "schema": { "type": "string" }, "description": "Award type" },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date from" },
          { "name": "dateTo", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Date to" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/intel/company-profile": {
      "get": {
        "operationId": "getCompanyProfile",
        "summary": "Get Company Competitive Profile",
        "description": "Get aggregated competitive intelligence profile (filings + news + contracts).",
        "tags": ["Intelligence"],
        "parameters": [
          { "name": "companyName", "in": "query", "schema": { "type": "string" }, "description": "Company name" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
...
  },
  "components": {
    "responses": {
      "SuccessResponse": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": true },
                "data": { "type": "object" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "RapidAPIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key"
      }
    }
  },
  "security": [
    { "RapidAPIKey": [] }
  ],
...
}